Github Samples - UE Blueprint Runtime Error
Summary:
We are getting a Blueprint Invalid Reference exception in MetaXRAudio unreal githud sample:
https://github.com/oculus-samples/Unreal-MetaXRAudioSDK

This diff alters the blueprint script and SignalDots actor to resolve this issue.
The primary issue was a missing ULoadnessNRT reference for audio magnitude analysis.

However, I did make some slight alterations to the blueprint logic to make it easier to read.
Also, we re-computed visibility for each skeletal mesh when it only needed to occur once.

Before:
 {F1976081347}
 {F1976081350}

After:
 {F1976081370}

Reviewed By: isaacmosebrook-meta

Differential Revision: D71215547

fbshipit-source-id: 11249b324b0ffe4532cc5ed62066f9c5feea0c06